make ObjId 16bit.
that should also be enough for now. we now can have 65536 objects per class

+++ src/util/BaseObject.h
@@ -23,7 +23,7 @@ struct AttrUpdateInfo {
 };
 
 struct BaseObject : DontCopyTag {
-	typedef uint32_t ObjId;
+	typedef uint16_t ObjId;
 	typedef ::WeakRef<BaseObject> WeakRef;
 
 	BaseObject();
